requirements.txt
gradio>=4.0.0
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
langchain>=1.0.0
langchain-core>=1.0.0
langchain-groq>=1.0.0
python-dotenv>=1.0.0
